project/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2
django-cache-url==3.4.4
django-redis==5.2.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.1
phonenumbers==8.13.11
psycopg2-binary==2.9.3
pycryptodome==3.14.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.4.0
